pyproject.toml
[project]
name = "school-mcp"
version = "0.1.0"
description = "MCP server for school management system backend"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "zhangwh", email = "1076787229@qq.com" }
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pyjwt>=2.8.0",
]

[project.scripts]
school-mcp = "school_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/school_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.20",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
